Added check for duplicate gradients. Closes #380 #329
