+++ test/unit/specs/Display.spec.js
@@ -1,23 +1,31 @@
-import { mount } from 'avoriaz';
-import Display from '@/components/Display';
+import Gradients from '../../../gradients.json';
 
-const wrapper = mount(Display, {
-  propsData: {
-    direction: 'to left',
-    gradient: {
-      colors: ['#1FA2FF', '#21DCF6', '#A6FFCB'],
-      name: 'Stripe',
-    },
-  },
-});
+const findDuplicatesinArrayByKeyValue = (array, propertyName) => {
+	let seenDuplicate = false;
+	let	testObject = {};
+	let dupes = [];
 
-describe('Display.vue', () => {
-  it('It renders element correctly', () => {
-    expect(wrapper.is('.display')).to.equal(true);
-  });
+	array.map((item) => {
+		var itemPropertyName = item[propertyName];
+		if (itemPropertyName in testObject) {
+			dupes.push(itemPropertyName);
+			testObject[itemPropertyName].duplicate = true;
+			item.duplicate = true;
+			seenDuplicate = true;
+		}
+		else {
+			testObject[itemPropertyName] = item;
+			delete item.duplicate;
+		}
+	})
 
-  it('should have a gradient as background style', () => {
-    expect(wrapper.hasStyle('background', 'linear-gradient(to left, #1FA2FF,#21DCF6,#A6FFCB)'))
-      .to.equal(true);
+	return seenDuplicate;
+}
+
+
+describe('Duplicate gradients', () => {
+  it('Has no duplicate gradient names', () => {
+    const dupes = findDuplicatesinArrayByKeyValue(Gradients, 'name');
+    expect(dupes).to.equal(false);
   });
 });
